fix: change monetary values to BigDecimal in DTOs

+++ src/main/java/com/vinisnzy/sales_management/dto/product/ProductRequestDTO.java
@@ -3,11 +3,13 @@ package com.vinisnzy.sales_management.dto.product;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record ProductRequestDTO(
         @NotBlank(message = "Name cannot be blank")
         String name,
 
         @Positive
-        Double price
+        BigDecimal price
 ) {
 }
