+++ src/main/java/com/vinisnzy/sales_management/dto/sale/SaleItemResponseDTO.java
@@ -1,9 +1,11 @@
 package com.vinisnzy.sales_management.dto.sale;
 
+import java.math.BigDecimal;
+
 public record SaleItemResponseDTO(
         Long id,
         String productName,
         Integer quantity,
-        Double subtotal
+        BigDecimal subtotal
 ) {
 }
